+++ tests/TestCase/Controller/Traits/RegisterTraitTest.php
@@ -14,6 +14,7 @@ namespace CakeDC\Users\Test\TestCase\Controller\Traits;
 use CakeDC\Users\Test\TestCase\Controller\Traits\BaseTraitTest;
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
+use Cake\Event\Event;
 use Cake\Mailer\Email;
 use Cake\ORM\TableRegistry;
 
@@ -88,6 +89,66 @@ class RegisterTraitTest extends BaseTraitTest
 
         $this->Trait->register();
 
+        $this->assertEquals(1, $this->table->find()->where(['username' => 'testRegistration'])->count());
+    }
+
+    /**
+     * Triggering beforeRegister event and not able to register the user
+     *
+     * @return void
+     */
+    public function testRegisterWithEventFalseResult()
+    {
+        $this->assertEquals(0, $this->table->find()->where(['username' => 'testRegistration'])->count());
+        $this->_mockRequestPost();
+        $this->_mockAuth();
+        $this->_mockFlash();
+        $this->_mockDispatchEvent(new Event('Users.Component.UsersAuth.beforeRegister'), ['username' => 'hello']);
+        $this->Trait->Flash->expects($this->once())
+            ->method('error')
+            ->with('The user could not be saved');
+        $this->Trait->expects($this->never())
+            ->method('redirect');
+        $this->Trait->request->expects($this->never())
+            ->method('is');
+
+        $this->Trait->register();
+        $this->assertEquals(0, $this->table->find()->where(['username' => 'testRegistration'])->count());
+    }
+
+    /**
+     * Triggering beforeRegister event and registering the user successfully
+     *
+     * @return void
+     */
+    public function testRegisterWithEventSuccessResult()
+    {
+        $data = [
+            'username' => 'testRegistration',
+            'password' => 'password',
+            'email' => 'test-registration@example.com',
+            'password_confirm' => 'password',
+            'tos' => 1
+        ];
+
+        $this->assertEquals(0, $this->table->find()->where(['username' => 'testRegistration'])->count());
+        $this->_mockRequestPost();
+        $this->_mockAuth();
+        $this->_mockFlash();
+        $this->_mockDispatchEvent(new Event('Users.Component.UsersAuth.beforeRegister'), $data);
+        $this->Trait->request->expects($this->once())
+            ->method('getData')
+            ->will($this->returnValue($data));
+        $this->Trait->Flash->expects($this->once())
+            ->method('success')
+            ->with('Please validate your account before log in');
+        $this->Trait->expects($this->once())
+            ->method('redirect')
+            ->with(['action' => 'login']);
+        $this->Trait->request->expects($this->never())
+            ->method('is');
+
+        $this->Trait->register();
         $this->assertEquals(1, $this->table->find()->where(['username' => 'testRegistration'])->count());
     }